==> 1_Import_of_data_from_database.ipynb <==
## Importing project data from a database using psycopg2 for the real estate EDA project

# Import the necessary libraries
import pandas as pd
import psycopg2

##### The database name, user name, passwort, etc. are safely storen (and "hidden") in a .env file in the same folder, so this information does not get shared. The os library is used to retrieve this information from the .env file for its use later on.

# Import os library and read in data from .env file
import os
from dotenv import load_dotenv

load_dotenv()
DATABASE = os.getenv('DATABASE')
USER_DB = os.getenv('USER_DB')
PASSWORD = os.getenv('PASSWORD')
HOST = os.getenv('HOST')
PORT = os.getenv('PORT')

##### pyscopg2 is now used to connect to the database.

# Create connection object conn
conn = psycopg2.connect(
    database=DATABASE,
    user=USER_DB,
    password=PASSWORD,
    host=HOST,
    port=PORT
)

##### A cursor is created to work inside the database.

cur = conn.cursor()

##### SQL commands can now be used to work inside the database while "being in the python environment".

# import the data into a pandas dataframe
# The database needs to know in which schma you want to work! This is set first.
command_sql1 = "SET SCHEMA 'eda'" 
cur.execute(command_sql1)
# The qury string joins the two tables of the database and with the pandas command a dataframe is created which consists of the two joined tables.
query_string = "SELECT * FROM eda.king_county_house_details d LEFT JOIN eda.king_county_house_sales s ON d.id = s.house_id"
df_psycopg = pd.read_sql(query_string, conn)

##### Closing the connection is important. Don't forget.

#close the connection
conn.close()

# Having a quick look at the imported data. Everything seems to be fine.
df_psycopg.head()

##### Data is locally stored inside a csv-file. This is not always an option, but for this project it is. Don't forget the "index=False" part, otherwise the data(frame) is expanded by this column.

#export the data to a csv-file
df_psycopg.to_csv('data/eda.csv',index=False)
