==> 2_Cleaning_of_real_estate_data.ipynb <==
## Cleaning of the imported data for the real estate EDA project

# import all libraries which can be helpful down the road
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import plotly.express as px
import altair as alt
import missingno as msno

##### Data needs to be imported to work with it. Then have a first look at the dataframe and the data itself.

# read data into a daframe, look at the first five rows to get an impression
df = pd.read_csv('data/eda.csv')
df.head()

# check types of columns, have a look if all columns posess values
df.info()

##### Some redundancies can be eliminated and some solumns can be renamed for a clearer description.

# house_id is the same as id, therefore the column can be dropped.
df.drop(columns=['house_id'], inplace=True)

# id.1 is a bad column name and it needs a better description anyways, rename id to prop(erty)_id to make it more obvious
df.rename(columns={"id.1": "transaction_id", "id": "prop_id"}, inplace=True)

# just have a look at the data
df.describe()

##### The data type of some columns can be changed, because there is no need to have floats as the datatype "everywhere". Additionally, NAs are dealt with in this part.
The removal of NAs and "0" values is done with consideration of the client. This procedure would probably vary if dealing with another client.  
For example, the date was given as a string and transformed into the pandas date format to make work easier.

# some columns are floats and this is not necessary - that's why I set them to integers
# three columns have NAs in them: if there is no basement, it cannot have an area (NA there --> 0 sqft). 
# If there is no value in the "view" column, I consider it as having no/(a bad) view -> set it to 0. If a property would have a good view, it would be noted for sure to increase the grade/value.
# if there is no value in the "waterfront" column, I consider it being a no water front lot --> set it to 0. If a property would have access to the water, it would highly likely be noted to increase the grade/value.

df[["sqft_living", "sqft_lot", "sqft_above", "sqft_living15", "sqft_lot15", "price"]] = df[["sqft_living", "sqft_lot", "sqft_above", "sqft_living15", "sqft_lot15", "price"]].astype(int)
df[["waterfront", "view", "sqft_basement"]] =  df[["waterfront", "view", "sqft_basement"]].fillna(0.0).astype(int)

# transforming the date column from a string to a date format
df['date'] = pd.to_datetime(df['date'], format='%Y-%m-%d')

# the yr_renovated column had an error in it. All years were inflated by factor 10. So I devided by 10 and then set the format to integer.
# All values with the value "0" were set to NA, because it's easier for plots and sorting. This datatype is not transferred, if you save the data in a csv-file. 
# You have to set it again after importing the data! It is a datatype used especially by pandas.

df["yr_renovated"] = df["yr_renovated"] / 10.0
df.loc[(df.yr_renovated == 0), 'yr_renovated'] = np.nan
df['yr_renovated'] = df['yr_renovated'].astype(pd.Int64Dtype())

# having a look at the dataframe after the first round of cleaning up the data.
df.info()

df.head()

##### After the cleaning of the data, in this case dropping columns, renaming columns, changing datatypes, dealing with NA and "0" values, the data is exported to a different csv-file. Please keep in mind, that this cannot be done everytime, because of space constraints for example.

# save cleaned date(frame) to a new csv-file
df.to_csv('data/cleaned_realestate.csv', sep=';', index=False)

##### So far the data has not been heavily edited or transformed. There has been no look at outliers or strange data points.   
##### Only the yr_renovated column was edited to gain sensible data for the year. The data was not created in the future ;)
