==> 4_Data_prep_client_Jacob_P.ipynb <==
## Adressing the clients needs, filter the data accordingly and derive proposals for the client

The client (Jacob Phillips) is a buyer and was described with the following paragraph:  
"Unlimited Budget, 4+ bathrooms or smaller house nearby, big lot (tennis court & pool), golf, historic, no waterfront"  

From this the following criteria were derived:
* money is not an issue --> no limitations
* number of bathrooms >= 4
* lot site > 20000 sqft (a tennis court needs 7200 sqft, a nicely sized pool needs 3000 sqft, house, garden, garage, pool house, walkways, etc.)
* historic house --> built before 1975
* no waterfront
  
The part "or smaller house nearby" was not addressed, because there were enough properties fulfilling the other criteria.  
The solution to the "golf wish" is documented in detail further below.

# import all libraries which can be helpful down the road
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import plotly.express as px
import altair as alt
import missingno as msno

# Data type has to be set again, because it is "lost" during saving and importing the data in/from csv-file
df = pd.read_csv('data/cleaned_realestate.csv', sep=";")
df['yr_renovated'] = df['yr_renovated'].astype(pd.Int64Dtype())
df.describe()

##### Creation of a dataframe with less data, only containing properties which fulfill the client's wishes. I added the criterium "condition >=4", because I assumed/decided that a client with those exculsive and expensive wishes would not want to move into a shabby house or spend much time renovating it before.

df_jacob_wide = df.query("bathrooms >=4 and yr_built < 1975 and waterfront == 0 and sqft_lot >= 20000 and condition >= 4")

df_jacob_wide

##### There was the need to manually remove two properties. Despite having the waterfront = 0 characteristic they were situated on the shoreline.

# prop_id 6072800170 is falsly annotated and is on the shoreline --> drop it
# prop_id 239000155 is falsly annotated, it lies on a lake --> drop it
df_jacob_wide.drop([7304], inplace=True)
df_jacob_wide.drop([18711], inplace=True)
df_jacob_wide

##### The locations of the ten remaining properties were shown on a map and this will be presented to the client.

fig = px.scatter_map(df_jacob_wide, 
                        lat="lat", 
                        lon="long", 
                        hover_name="prop_id", 
                        color_discrete_sequence=['blue'],
                        zoom=9, 
                        height=800,
                        width=800)

fig.update_traces(marker=dict(size=9),
                  selector=dict(mode='markers'))

fig.update_layout(mapbox_style="open-street-map")
fig.update_layout(margin={"r":0,"t":0,"l":0,"b":0})
fig.show()

##### With the help of two websites, I identified 19 private(!) golf courses in the Seattle area and converted their addresses to GPS coordinates. This data was combined in a csv-file, so that the data could be readily accessed.
---
##### Then the dataframe of the ten remaining properties was combined with the dataframe of the golf courses. Then the distance between the properties and all golf courses was calculated (linear distance, not road distance - this is a drawback!). Finally, the distances between a property and its five closest golf courses was averaged and used as the final ranking criterium.  
---  
The formula to calculate the distances was adapted from stack overflow. By now there is a library (H3?) which can be used with pandas, which would do the same. A few distances were checked for accuracy by Google maps.  

---
But first a dataframe was created which consisted of the combined locations of the properties and the golf courses, so that the figure can be shown to the client.


df_golf_courses_temp = pd.read_csv('data/golf_clubs.csv', sep=";")
df_golf_courses_temp.rename(columns={'g_lat': 'lat', 'g_long': 'long'}, inplace=True)
df_dist_golf_plot = pd.concat([df_jacob_wide_shrunk, df_golf_courses_temp], ignore_index=True, sort=False)
df_dist_golf_plot.loc[df_dist_golf_plot.prop_id > 0, 'Location'] = "Property"
df_dist_golf_plot.loc[df_dist_golf_plot.prop_id.isna(), 'Location'] = "Golf course"

fig = px.scatter_map(df_dist_golf_plot, 
                        lat="lat", 
                        lon="long", 
                        hover_name="prop_id", 
                        #hover_data=["Address", "Listed"],
                        color="Location",
                        #color_continuous_scale=color_scale,
                        #size="Listed",
                        zoom=9, 
                        height=800,
                        width=900)

fig.update_traces(marker=dict(size=9),
                  selector=dict(mode='markers'))

fig.update_layout(mapbox_style="open-street-map")
fig.update_layout(margin={"r":0,"t":0,"l":0,"b":0})
fig.show()

# import of golf data and combining it with real estate data
df_golf_courses = pd.read_csv('data/golf_clubs.csv', sep=";")
df_jacob_wide_shrunk = df_jacob_wide[['prop_id', 'lat', 'long']] 
df_dist_golf = pd.merge(df_jacob_wide_shrunk, df_golf_courses, how="cross")

# function to calculate the distances
def haversine(lat1, lon1, lat2, lon2):
    """
    Calculate the great circle distance between two points
    on the earth (specified in decimal degrees)
    """
    # Convert latitude and longitude from degrees to radians
    lat1, lon1, lat2, lon2 = map(np.radians, [lat1, lon1, lat2, lon2])

    # Haversine formula
    dlon = lon2 - lon1
    dlat = lat2 - lat1
    a = np.sin(dlat/2)**2 + np.cos(lat1) * np.cos(lat2) * np.sin(dlon/2)**2
    c = 2 * np.arcsin(np.sqrt(a))
    r = 6371  # Radius of earth in kilometers. Use 3956 for miles
    return c * r

df_dist_golf['distance'] = df_dist_golf.apply(lambda row: haversine(row['lat'], row['long'], row['g_lat'], row['g_long']), axis=1)
# sort the distances distances per property and select the five shortest of them
df_dist_golf = df_dist_golf.sort_values('distance').groupby('prop_id').head(5).sort_values(['prop_id', 'distance'])
# calculate the average of the five distances
df_dist_golf = df_dist_golf.groupby(['prop_id'])['distance'].mean().reset_index(name='g_average_distance')
# join the property and the distance data
df_jacob_wide_golf = pd.merge(df_jacob_wide, df_dist_golf, on='prop_id')
df_dist_golf.head(15)
# select the top five properties (by distance to the golf courses)
df_jacob_wide_golf.sort_values('g_average_distance')[:5]

##### Finally, the coordinates of the top3 properties were entered into Google maps, to create pictures from high above (to see where the property is situated) and from closer range to get provide an overview how the property looks. If there were good pictures available from Google streetview those were added as well.  
---
##### Data from the dataframe was combined with the Google images to present the top3 properties to the client. Those things constitute the final slides of the presentation. Now, one can only hope he likes the proposals ;)
